Fix cast in unit tests

+++ src/gui/shellwidget/test/test_cell.cpp
@@ -12,7 +12,7 @@ class Test: public QObject
 private slots:
 	void cellDefault() {
 		Cell c;
-		QCOMPARE(c.c, ' ');
+		QCOMPARE(c.c, uint(' '));
 		// Default colors are invalid
 		QCOMPARE(c.foregroundColor, QColor());
 		QCOMPARE(c.backgroundColor, QColor());
